1.5.20152.50
Updates to support Xamarin.Forms 1.5

+++ FlexPie101/FlexPie101.Data/Model/FruitEntity.cs
@@ -10,9 +10,9 @@ namespace FlexPieDemo.Data.Model
     {
         public string Name { get; set; }
 
-        public decimal Value { get; set; }
+        public double Value { get; set; }
 
-        public FruitEntity(string name, decimal value)
+        public FruitEntity(string name, double value)
         {
             this.Name = name;
             this.Value = value;
